Cohen's w (omega)

Calculates Cohen's w for a table of nominal variables.

cohenW(x, y = NULL, p = NULL, digits = 4, ...)

Arguments

x
Either a two-way table or a two-way matrix. Can also be a vector of observations for one dimension of a two-way table.

y
If x is a vector, y is the vector of observations for the second dimension of a two-way table.

p
If x is a vector of observed counts, p can be given as a vector of expected probabilties, as in a chi-square goodness of fit test.

digits
The number of significant digits in the output.

...
Additional arguments passed to chisq.test.

Value

A single statistic, Cohen's w.

Details

Cohen's w is used as a measure of association between two nominal variables, or as an effect size for a chi-square test of association. For a 2 x 2 table, the absolute value of the phi statistic is the same as Cohen's w. The value of Cohen's w is not bound by 1 on the upper end.

References

http://rcompanion.org/handbook/H_10.html

See also

Examples

### Example with table
data(Anderson)
fisher.test(Anderson)

#> 
#> 	Fisher's Exact Test for Count Data
#> 
#> data:  Anderson
#> p-value = 0.000668
#> alternative hypothesis: two.sided
#> 

cohenW(Anderson)

#> Cohen w 
#>  0.4387 

### Example for goodness-of-fit
### Bird foraging example, Handbook of Biological Statistics
observed = c(70,   79,   3,    4)
expected = c(0.54, 0.40, 0.05, 0.01)
chisq.test(observed, p = expected)

#> Warning: Chi-squared approximation may be incorrect

#> 
#> 	Chi-squared test for given probabilities
#> 
#> data:  observed
#> X-squared = 13.593, df = 3, p-value = 0.003514
#> 

cohenW(observed, p = expected)

#> Cohen w 
#>  0.2952 

### Example with two vectors
Species = c(rep("Species1", 16), rep("Species2", 16))
Color   = c(rep(c("blue", "blue", "blue", "green"),4),
            rep(c("green", "green", "green", "blue"),4))
fisher.test(Species, Color)

#> 
#> 	Fisher's Exact Test for Count Data
#> 
#> data:  Species and Color
#> p-value = 0.01211
#> alternative hypothesis: true odds ratio is not equal to 1
#> 95 percent confidence interval:
#>   1.463598 60.596055
#> sample estimates:
#> odds ratio 
#>   8.279362 
#> 

cohenW(Species, Color)

#> Cohen w 
#>     0.5