Compare fit statistics for lm models

Produces a table of fit statistics for multiple lm models.

compareLM(fits, ...)

Arguments

fits
A series of model object names, separated by commas.

...
Other arguments passed to list.

Value

A list of two objects: The series of model calls, and a data frame of statistics for each model.

Details

Produces a table of fit statistics for multiple lm models: AIC, AICc, BIC, p-value, R-squared, and adjusted R-squared.

Smaller values for AIC, AICc, and BIC indicate a better balance of goodness-of-fit of the model and the complexity of the model. The goal is to find a model that adequately explains the data without having too many terms.

BIC tends to choose models with fewer parameters relative to AIC.

In the table, Shapiro.W and Shapiro.p are the W statistic and p-value for the Shapiro-Wilks test on the residuals of the model.

For comparisons with AIC, etc., to be valid, both models must have the same data, without transformations, use the same dependent variable, and be fit with the same method. They do not need to be nested.

References

http://rcompanion.org/handbook/I_10.html http://rcompanion.org/rcompanion/e_05.html

See also

Examples

### Compare among polynomial models
data(BrendonSmall)
BrendonSmall$Calories = as.numeric(BrendonSmall$Calories)

BrendonSmall$Calories2 = BrendonSmall$Calories * BrendonSmall$Calories
BrendonSmall$Calories3 = BrendonSmall$Calories * BrendonSmall$Calories *
                         BrendonSmall$Calories
BrendonSmall$Calories4 = BrendonSmall$Calories * BrendonSmall$Calories *
                         BrendonSmall$Calories * BrendonSmall$Calories
model.1 = lm(Sodium ~ Calories, data = BrendonSmall)
model.2 = lm(Sodium ~ Calories + Calories2, data = BrendonSmall)
model.3 = lm(Sodium ~ Calories + Calories2 + Calories3, data = BrendonSmall)
model.4 = lm(Sodium ~ Calories + Calories2 + Calories3 + Calories4,
             data = BrendonSmall)
compareLM(model.1, model.2, model.3, model.4)

#> $Models
#>   Formula                                                
#> 1 "Sodium ~ Calories"                                    
#> 2 "Sodium ~ Calories + Calories2"                        
#> 3 "Sodium ~ Calories + Calories2 + Calories3"            
#> 4 "Sodium ~ Calories + Calories2 + Calories3 + Calories4"
#> 
#> $Fit.criteria
#>   Rank Df.res   AIC  AICc   BIC R.squared Adj.R.sq   p.value Shapiro.W
#> 1    2     43 461.1 461.7 466.5    0.7207   0.7142 1.737e-13    0.9696
#> 2    3     42 431.4 432.4 438.6    0.8621   0.8556 8.487e-19    0.9791
#> 3    4     41 433.2 434.7 442.2    0.8627   0.8526 1.025e-17    0.9749
#> 4    5     40 428.5 430.8 439.4    0.8815   0.8696 5.566e-18    0.9625
#>   Shapiro.p
#> 1    0.2800
#> 2    0.5839
#> 3    0.4288
#> 4    0.1518
#>