Cramer's V (phi)

Calculates Cramer's V for a table of nominal variables; confidence intervals by bootstrap.

cramerV(x, y = NULL, ci = FALSE, conf = 0.95, type = "perc",
  R = 1000, histogram = FALSE, digits = 4, bias.correct = FALSE,
  ...)

Arguments

x
Either a two-way table or a two-way matrix. Can also be a vector of observations for one dimension of a two-way table.

y
If x is a vector, y is the vector of observations for the second dimension of a two-way table.

ci
If TRUE, returns confidence intervals by bootstrap. May be slow.

conf
The level for the confidence interval.

type
The type of confidence interval to use. Can be any of "norm", "basic", "perc", or "bca". Passed to boot.ci.

R
The number of replications to use for bootstrap.

histogram
If TRUE, produces a histogram of bootstrapped values.

digits
The number of significant digits in the output.

bias.correct
If TRUE, a bias correction is applied.

...
Additional arguments passed to chisq.test.

Value

A single statistic, Cramer's V. Or a small data frame consisting of Cramer's V, and the lower and upper confidence limits.

Details

Cramer's V is used as a measure of association between two nominal variables, or as an effect size for a chi-square test of association. For a 2 x 2 table, the absolute value of the phi statistic is the same as Cramer's V.

When V is close to 0 or very large, or with small counts, the confidence intervals determined by this method may not be reliable, or the procedure may fail.

References

http://rcompanion.org/handbook/H_10.html

See also

Examples

### Example with table
data(Anderson)
fisher.test(Anderson)

#> 
#> 	Fisher's Exact Test for Count Data
#> 
#> data:  Anderson
#> p-value = 0.000668
#> alternative hypothesis: two.sided
#> 

cramerV(Anderson)

#> Cramer V 
#>   0.4387 

### Example with two vectors
Species = c(rep("Species1", 16), rep("Species2", 16))
Color   = c(rep(c("blue", "blue", "blue", "green"),4),
            rep(c("green", "green", "green", "blue"),4))
fisher.test(Species, Color)

#> 
#> 	Fisher's Exact Test for Count Data
#> 
#> data:  Species and Color
#> p-value = 0.01211
#> alternative hypothesis: true odds ratio is not equal to 1
#> 95 percent confidence interval:
#>   1.463598 60.596055
#> sample estimates:
#> odds ratio 
#>   8.279362 
#> 

cramerV(Species, Color)

#> Cramer V 
#>      0.5