Groupwise means and confidence intervals

Calculates means and confidence intervals for groups.

groupwiseMean(
  formula = NULL,
  data = NULL,
  var = NULL,
  group = NULL,
  conf = 0.95,
  R = 5000,
  boot = FALSE,
  traditional = TRUE,
  normal = FALSE,
  basic = FALSE,
  percentile = FALSE,
  bca = FALSE,
  digits = 3,
  ...
)

Arguments

formula
A formula indicating the measurement variable and the grouping variables. e.g. y ~ x1 + x2.

data
The data frame to use.

var
The measurement variable to use. The name is in double quotes.

group
The grouping variable to use. The name is in double quotes. Multiple names are listed as a vector. (See example.)

conf
The confidence interval to use.

R
The number of bootstrap replicates to use for bootstrapped statistics.

boot
If TRUE, includes the mean of the bootstrapped means. This can be used as an estimate of the mean for the group.

traditional
If TRUE, includes the traditional confidence intervals for the group means, using the t-distribution.

normal
If TRUE, includes the normal confidence intervals for the group means by bootstrap. See boot.ci.

basic
If TRUE, includes the basic confidence intervals for the group means by bootstrap. See boot.ci.

percentile
If TRUE, includes the percentile confidence intervals for the group means by bootstrap. See boot.ci.

bca
If TRUE, includes the BCa confidence intervals for the group means by bootstrap. See boot.ci.

digits
The number of significant figures to use in output.

...
Other arguments passed to the boot function.

Value

A data frame of requested statistics by group.

Details

The input should include either formula and data; or data, var, and group. (See examples).

With some options, the function may not handle missing values well. This seems to happen particularly with bca = TRUE.

Note

The parsing of the formula is simplistic. The first variable on the left side is used as the measurement variable. The variables on the right side are used for the grouping variables.

Results for ungrouped (one-sample) data can be obtained by either setting the right side of the formula to 1, e.g. y ~ 1, or by setting group=NULL.

References

http://rcompanion.org/handbook/C_03.html

See also

Examples

### Example with formula notation
data(Catbus)
groupwiseMean(Steps ~ Teacher + Sex,
              data        = Catbus,
              traditional = FALSE,
              percentile  = TRUE)

#>   Teacher    Sex n Mean Conf.level Percentile.lower Percentile.upper
#> 1  Catbus female 6 8000       0.95             7000             9000
#> 2  Catbus   male 4 7000       0.95             5500             8500
#> 3 Satsuki female 4 8500       0.95             8000             9000
#> 4 Satsuki   male 3 7000       0.95             6000             8000
#> 5  Totoro female 5 8200       0.95             7000             9400
#> 6  Totoro   male 4 7000       0.95             6250             7750

### Example with variable notation
data(Catbus)
groupwiseMean(data        = Catbus,
              var         = "Steps",
              group       = c("Teacher", "Sex"),
              traditional = FALSE,
              percentile  = TRUE)

#>   Teacher    Sex n Mean Conf.level Percentile.lower Percentile.upper
#> 1  Catbus female 6 8000       0.95             7000             9000
#> 2  Catbus   male 4 7000       0.95             5500             8500
#> 3 Satsuki female 4 8500       0.95             8000             9000
#> 4 Satsuki   male 3 7000       0.95             6000             8000
#> 5  Totoro female 5 8200       0.95             7000             9400
#> 6  Totoro   male 4 7000       0.95             6250             7750