Groupwise medians and confidence intervals

Calculates medians and confidence intervals for groups.

groupwiseMedian(
  formula = NULL,
  data = NULL,
  var = NULL,
  group = NULL,
  conf = 0.95,
  R = 5000,
  boot = FALSE,
  pseudo = FALSE,
  basic = FALSE,
  normal = FALSE,
  percentile = FALSE,
  bca = TRUE,
  wilcox = FALSE,
  exact = FALSE,
  digits = 3,
  ...
)

Arguments

formula
A formula indicating the measurement variable and the grouping variables. e.g. y ~ x1 + x2.

data
The data frame to use.

var
The measurement variable to use. The name is in double quotes.

group
The grouping variable to use. The name is in double quotes. Multiple names are listed as a vector. (See example.)

conf
The confidence interval to use.

R
The number of bootstrap replicates to use for bootstrapped statistics.

boot
If TRUE, includes the mean of the bootstrapped medians. This can be used as an estimate of the median for the group.

pseudo
If TRUE, includes the pseudo median from wilcox.test.

basic
If TRUE, includes the basic confidence intervals for the group means by bootstrap. See boot::boot.ci.

normal
If TRUE, includes the normal confidence intervals for the group means by bootstrap. See boot::boot.ci.

percentile
If TRUE, includes the percentile confidence intervals for the group means by bootstrap. See boot::boot.ci.

bca
If TRUE, includes the BCa confidence intervals for the group means by bootstrap. See boot::boot.ci.

wilcox
If TRUE, includes the wilcox confidence intervals from stats::wilcox.test.

exact
If TRUE, includes the "exact" confidence intervals from DescTools::MedianCI.

digits
The number of significant figures to use in output.

...
Other arguments passed to the boot function.

Value

A data frame of requested statistics by group.

Details

The input should include either formula and data; or data, var, and group. (See examples).

With some options, the function may not handle missing values well. This seems to happen particularly with bca = TRUE.

Note

The parsing of the formula is simplistic. The first variable on the left side is used as the measurement variable. The variables on the right side are used for the grouping variables.

Results for ungrouped (one-sample) data can be obtained by either setting the right side of the formula to 1, e.g. y ~ 1, or by setting group=NULL.

References

http://rcompanion.org/handbook/E_04.html

See also

Author

Salvatore Mangiafico, mangiafico@njaes.rutgers.edu

Examples

### Example with formula notation
data(Catbus)
groupwiseMedian(Steps ~ Teacher + Sex,
                data        = Catbus,
                bca         = FALSE,
                percentile  = TRUE,
                R           = 1000)

#>   Teacher    Sex n Median Conf.level Percentile.lower Percentile.upper
#> 1  Catbus female 6   8000       0.95             6500             9500
#> 2  Catbus   male 4   7000       0.95             5000             9000
#> 3 Satsuki female 4   8500       0.95             8000             9000
#> 4 Satsuki   male 3   7000       0.95             6000             8000
#> 5  Totoro female 5   8000       0.95             6000            10000
#> 6  Totoro   male 4   7000       0.95             6000             8000

                
### Example with variable notation
data(Catbus)
groupwiseMedian(data         = Catbus,
                var         = "Steps",
                group       = c("Teacher", "Sex"),
                bca         = FALSE,
                percentile  = TRUE,
                R           = 1000)

#>   Teacher    Sex n Median Conf.level Percentile.lower Percentile.upper
#> 1  Catbus female 6   8000       0.95             6500             9500
#> 2  Catbus   male 4   7000       0.95             5000             9000
#> 3 Satsuki female 4   8500       0.95             8000             9000
#> 4 Satsuki   male 3   7000       0.95             6000             8000
#> 5  Totoro female 5   8000       0.95             6000            10000
#> 6  Totoro   male 4   7000       0.95             6000             8000