Pairwise Mood's median tests

Conducts pairwise Mood's median tests across groups.

pairwiseMedianTest(
  formula = NULL,
  data = NULL,
  x = NULL,
  g = NULL,
  digits = 4,
  method = "fdr",
  ...
)

Arguments

formula
A formula indicating the measurement variable and the grouping variable. e.g. y ~ group.

data
The data frame to use.

x
The response variable as a vector.

g
The grouping variable as a vector.

digits
The number of significant digits to round output.

method
The p-value adjustment method to use for multiple tests. See stats::p.adjust.

...
Additional arguments passed to codecoin::median_test.

Value

A dataframe of the groups being compared, the p-values, and the adjusted p-values.

Details

The input should include either formula and data; or x, and g.

Mood's median test compares medians among two or more groups. See http://rcompanion.org/handbook/F_09.html for futher discussion of this test.

The pairwiseMedianTest function can be used as a post-hoc method following an omnibus Mood's median test.

Note

The parsing of the formula is simplistic. The first variable on the left side is used as the measurement variable. The first variable on the right side is used for the grouping variable.

References

http://rcompanion.org/handbook/F_09.html

See also

Examples

data(PoohPiglet)
PoohPiglet$Speaker = factor(PoohPiglet$Speaker,
                     levels = c("Pooh", "Tigger", "Piglet"))
PT = pairwiseMedianTest(Likert ~ Speaker,
                        data   = PoohPiglet,
                        exact  = NULL,
                        method = "fdr")

#>    x      g     g2
#> 1  3   Pooh   Pooh
#> 2  5   Pooh   Pooh
#> 3  4   Pooh   Pooh
#> 4  4   Pooh   Pooh
#> 5  4   Pooh   Pooh
#> 6  4   Pooh   Pooh
#> 7  4   Pooh   Pooh
#> 8  4   Pooh   Pooh
#> 9  5   Pooh   Pooh
#> 10 5   Pooh   Pooh
#> 21 4 Tigger Tigger
#> 22 4 Tigger Tigger
#> 23 4 Tigger Tigger
#> 24 4 Tigger Tigger
#> 25 5 Tigger Tigger
#> 26 3 Tigger Tigger
#> 27 5 Tigger Tigger
#> 28 4 Tigger Tigger
#> 29 4 Tigger Tigger
#> 30 3 Tigger Tigger
#>    x      g     g2
#> 1  3   Pooh   Pooh
#> 2  5   Pooh   Pooh
#> 3  4   Pooh   Pooh
#> 4  4   Pooh   Pooh
#> 5  4   Pooh   Pooh
#> 6  4   Pooh   Pooh
#> 7  4   Pooh   Pooh
#> 8  4   Pooh   Pooh
#> 9  5   Pooh   Pooh
#> 10 5   Pooh   Pooh
#> 11 2 Piglet Piglet
#> 12 4 Piglet Piglet
#> 13 2 Piglet Piglet
#> 14 2 Piglet Piglet
#> 15 1 Piglet Piglet
#> 16 2 Piglet Piglet
#> 17 3 Piglet Piglet
#> 18 2 Piglet Piglet
#> 19 2 Piglet Piglet
#> 20 3 Piglet Piglet
#>    x      g     g2
#> 21 4 Tigger Tigger
#> 22 4 Tigger Tigger
#> 23 4 Tigger Tigger
#> 24 4 Tigger Tigger
#> 25 5 Tigger Tigger
#> 26 3 Tigger Tigger
#> 27 5 Tigger Tigger
#> 28 4 Tigger Tigger
#> 29 4 Tigger Tigger
#> 30 3 Tigger Tigger
#> 11 2 Piglet Piglet
#> 12 4 Piglet Piglet
#> 13 2 Piglet Piglet
#> 14 2 Piglet Piglet
#> 15 1 Piglet Piglet
#> 16 2 Piglet Piglet
#> 17 3 Piglet Piglet
#> 18 2 Piglet Piglet
#> 19 2 Piglet Piglet
#> 20 3 Piglet Piglet

PT

#>            Comparison   p.value p.adjust
#> 1   Pooh - Tigger = 0    0.6147 0.614700
#> 2   Pooh - Piglet = 0 0.0004883 0.001465
#> 3 Tigger - Piglet = 0  0.002165 0.003247

cldList(comparison = PT$Comparison,
        p.value    = PT$p.adjust,
        threshold  = 0.05)

#>    Group Letter MonoLetter
#> 1   Pooh      a         a 
#> 2 Tigger      a         a 
#> 3 Piglet      b          b