Pairwise two-sample permutation symmetry tests

Conducts pairwise two-sample permutation tests of symmetry across groups.

pairwisePermutationSymmetry(formula = NULL, data = NULL, x = NULL,
  g = NULL, b = NULL, method = "fdr", ...)

Arguments

formula
A formula indicating the measurement variable and the grouping variable. e.g. y ~ group | block.

data
The data frame to use.

x
The response variable as a vector.

g
The grouping variable as a vector.

b
The blocking variable as a vector.

method
The p-value adjustment method to use for multiple tests. See stats::p.adjust.

...
Additional arguments passed to coin::symmetry_test.

Value

A dataframe of the groups being compared, the p-values, and the adjusted p-values.

Details

The input should include either formula and data; or x, g, and b.

Permutation tests are non-parametric tests that do not assume normally-distributed errors. See http://rcompanion.org/rcompanion/d_06a.html for futher discussion of this test.

The pairwisePermutationSymmetry function can be used as a post-hoc method following an omnibus permutation test analogous to a paired one-way analysis of variance.

Note

The parsing of the formula is simplistic. The first variable on the left side is used as the measurement variable. The first variable on the right side is used for the grouping variable. The second variable on the right side is used for the blocking variable.

References

http://rcompanion.org/handbook/K_03.html

See also

Examples

data(BobBelcher)
BobBelcher$Instructor = factor( BobBelcher$Instructor,
                  levels = c("Linda Belcher", "Louise Belcher",
                             "Tina Belcher", "Bob Belcher",
                             "Gene Belcher"))
BobBelcher$Likert.f = factor(BobBelcher$Likert, ordered=TRUE)
PT = pairwisePermutationSymmetry(Likert.f ~ Instructor | Rater,
                                 data      = BobBelcher,
                                 method = "fdr")
PT

#>                            Comparison    Stat  p.value p.adjust
#> 1  Linda Belcher - Louise Belcher = 0 -0.9045   0.3657  0.42680
#> 2    Linda Belcher - Tina Belcher = 0   0.378   0.7055  0.70550
#> 3     Linda Belcher - Bob Belcher = 0    2.38  0.01729  0.03458
#> 4    Linda Belcher - Gene Belcher = 0   2.593 0.009522  0.03458
#> 5   Louise Belcher - Tina Belcher = 0   1.155   0.2482  0.35460
#> 6    Louise Belcher - Bob Belcher = 0   2.265  0.02354  0.03923
#> 7   Louise Belcher - Gene Belcher = 0   2.744 0.006068  0.03458
#> 8      Tina Belcher - Bob Belcher = 0   2.412  0.01586  0.03458
#> 9     Tina Belcher - Gene Belcher = 0   2.528  0.01147  0.03458
#> 10     Bob Belcher - Gene Belcher = 0  0.8704   0.3841  0.42680

cldList(comparison = PT$Comparison,
        p.value    = PT$p.adjust,
        threshold  = 0.05)

#>           Group Letter MonoLetter
#> 1  LindaBelcher      a         a 
#> 2 LouiseBelcher      a         a 
#> 3   TinaBelcher      a         a 
#> 4    BobBelcher      b          b
#> 5   GeneBelcher      b          b