Pairwise two-sample permutation symmetry tests with matrix output

Conducts pairwise two-sample permutation tests for symmetry across groups.

pairwisePermutationSymmetryMatrix(
  formula = NULL,
  data = NULL,
  x = NULL,
  g = NULL,
  b = NULL,
  method = "fdr",
  ...
)

Arguments

formula
A formula indicating the measurement variable and the grouping variable. e.g. y ~ group.

data
The data frame to use.

x
The response variable as a vector.

g
The grouping variable as a vector.

b
The blocking variable as a vector.

method
The p-value adjustment method to use for multiple tests. See stats::p.adjust.

...
Additional arguments passed to coin::symmetry_test.

Value

A list consisting of: A matrix of p-values; the p-value adjustment method; a matrix of adjusted p-values.

Details

The input should include either formula and data; or x, g, and b.

Permutation tests are non-parametric tests that do not assume normally-distributed errors. See http://rcompanion.org/rcompanion/d_06a.html for futher discussion of this test.

The pairwisePermutationSymmetryMatrix function can be used as a post-hoc method following an omnibus permutation test analogous to a paired one-way analysis of variance. The matrix output can be converted to a compact letter display.

Note

The parsing of the formula is simplistic. The first variable on the left side is used as the measurement variable. The first variable on the right side is used for the grouping variable. The second variable on the right side is used for the blocking variable.

References

http://rcompanion.org/handbook/K_03.html

See also

Author

Salvatore Mangiafico, mangiafico@njaes.rutgers.edu

Examples

data(BobBelcher)
BobBelcher$Instructor = factor( BobBelcher$Instructor, 
                  levels = c("Linda Belcher", "Louise Belcher",
                             "Tina Belcher", "Bob Belcher",
                             "Gene Belcher"))
BobBelcher$Likert.f = factor(BobBelcher$Likert, ordered=TRUE)
PT = pairwisePermutationSymmetryMatrix(Likert.f ~ Instructor | Rater,
                                       data   = BobBelcher,
                                       method = "fdr")$Adjusted
PT

#>                Linda Belcher Louise Belcher Tina Belcher Bob Belcher
#> Linda Belcher        1.00000        0.42680      0.70550     0.03458
#> Louise Belcher       0.42680        1.00000      0.35460     0.03923
#> Tina Belcher         0.70550        0.35460      1.00000     0.03458
#> Bob Belcher          0.03458        0.03923      0.03458     1.00000
#> Gene Belcher         0.03458        0.03458      0.03458     0.42680
#>                Gene Belcher
#> Linda Belcher       0.03458
#> Louise Belcher      0.03458
#> Tina Belcher        0.03458
#> Bob Belcher         0.42680
#> Gene Belcher        1.00000

library(multcompView)
multcompLetters(PT,
                compare="<",
                threshold=0.05,
                Letters=letters)   

#>  Linda Belcher Louise Belcher   Tina Belcher    Bob Belcher   Gene Belcher 
#>            "a"            "a"            "a"            "b"            "b"