Pairwise two-sample permutation tests

Conducts pairwise two-sample permutation tests across groups.

pairwisePermutationTest(
  formula = NULL,
  data = NULL,
  x = NULL,
  g = NULL,
  method = "fdr",
  ...
)

Arguments

formula
A formula indicating the measurement variable and the grouping variable. e.g. y ~ group.

data
The data frame to use.

x
The response variable as a vector.

g
The grouping variable as a vector.

method
The p-value adjustment method to use for multiple tests. See stats::p.adjust.

...
Additional arguments passed to coin::independence_test.

Value

A dataframe of the groups being compared, the p-values, and the adjusted p-values.

Details

The input should include either formula and data; or x, and g.

Permutation tests are non-parametric tests that do not assume normally-distributed errors. See http://rcompanion.org/rcompanion/d_06a.html for futher discussion of this test.

The pairwisePermutationTest function can be used as a post-hoc method following an omnibus permutation test analogous to a one-way analysis of variance.

Note

The parsing of the formula is simplistic. The first variable on the left side is used as the measurement variable. The first variable on the right side is used for the grouping variable.

References

http://rcompanion.org/handbook/K_02.html

See also

Author

Salvatore Mangiafico, mangiafico@njaes.rutgers.edu

Examples

data(PoohPiglet)
PoohPiglet$Speaker = factor(PoohPiglet$Speaker,
                     levels = c("Pooh", "Tigger", "Piglet"))
PT = pairwisePermutationTest(Likert ~   Speaker,
                             data   = PoohPiglet,
                             method = "fdr")
PT

#>            Comparison  Stat   p.value p.adjust
#> 1   Pooh - Tigger = 0 0.698    0.4852 0.485200
#> 2   Pooh - Piglet = 0 3.515  0.000439 0.001238
#> 3 Tigger - Piglet = 0 3.344 0.0008254 0.001238

cldList(comparison = PT$Comparison,
        p.value    = PT$p.adjust,
        threshold  = 0.05)

#>    Group Letter MonoLetter
#> 1   Pooh      a         a 
#> 2 Tigger      a         a 
#> 3 Piglet      b          b