r effect size for Wilcoxon one-sample signed-rank test

Calculates r effect size for a Wilcoxon one-sample signed-rank test; confidence intervals by bootstrap.

wilcoxonOneSampleR(
  x,
  mu = NULL,
  coin = FALSE,
  ci = FALSE,
  conf = 0.95,
  type = "perc",
  R = 1000,
  histogram = FALSE,
  digits = 3,
  ...
)

Arguments

x
A vector of observations.

mu
The value to compare x to, as in wilcox.test

coin
If FALSE, the default, the Z value is extracted from a function similar to the wilcox.test function in the stats package. If TRUE, the Z value is extracted from the wilcox_test function in the coin package. This method may be much slower, especially if a confidence interval is produced.

ci
If TRUE, returns confidence intervals by bootstrap. May be slow.

conf
The level for the confidence interval.

type
The type of confidence interval to use. Can be any of "norm", "basic", "perc", or "bca". Passed to boot.ci.

R
The number of replications to use for bootstrap.

histogram
If TRUE, produces a histogram of bootstrapped values.

digits
The number of significant digits in the output.

...
Additional arguments passed to the wilcoxsign_test function.

Value

A single statistic, r. Or a small data frame consisting of r, and the lower and upper confidence limits.

Details

r is calculated as Z divided by square root of the number of observations.

The calculated statistic is equivalent to the statistic returned by the wilcoxPairedR function with one group equal to a vector of mu. The author knows of no reference for this technique.

Currently, the function makes no provisions for NA values in the data. It is recommended that NAs be removed beforehand.

When the data are greater than mu, r is positive. When the data are less than mu, r is negative.

When r is close to extremes, or with small counts in some cells, the confidence intervals determined by this method may not be reliable, or the procedure may fail.

References

http://rcompanion.org/handbook/F_02.html

Author

Salvatore Mangiafico, mangiafico@njaes.rutgers.edu

Examples

data(Pooh)
Data = Pooh[Pooh$Time==2,]
wilcox.test(Data$Likert, mu=3, exact=FALSE)

#> 
#> 	Wilcoxon signed rank test with continuity correction
#> 
#> data:  Data$Likert
#> V = 36, p-value = 0.01154
#> alternative hypothesis: true location is not equal to 3
#> 

wilcoxonOneSampleR(x = Data$Likert, mu=3)

#>     r 
#> 0.822