Cohen's g and odds ratio for paired contingency tables

Calculates Cohen's g and odds ratio for paired contingency tables, such as those that might be analyzed with McNemar or McNemar-Bowker tests.

cohenG(x, ci = FALSE, conf = 0.95, type = "perc", R = 1000,
  histogram = FALSE, digits = 3)

Arguments

x
A two-way contingency table. It must be square. It can have two or more levels for each dimension.

ci
If TRUE, returns confidence intervals by bootstrap. May be slow.

conf
The level for the confidence interval.

type
The type of confidence interval to use. Can be any of "norm", "basic", "perc", or "bca". Passed to boot.ci.

R
The number of replications to use for bootstrap.

histogram
If TRUE, produces a histogram of bootstrapped values.

digits
The number of significant digits in the output.

Value

A list containing: a data frame of results of the global statistics; and a data frame of results of the pairwise statistics.

Details

For a 2 x 2 table, where a and d are the concordant cells and b and c are discordant cells: Odds ratio is the greater of b/c or c/b; P is the greater of b/(b+c) or c/(b+c); and Cohen's g is P - 0.5. These statistics are extended to tables larger than 2 x 2.

When the reported statistics are close to their extremes, or with small counts, the confidence intervals determined by this method may not be reliable, or the procedure may fail.

References

http://rcompanion.org/handbook/H_05.html

See also

Examples

### 2 x 2 repeated matrix example
data(AndersonRainBarrel)
cohenG(AndersonRainBarrel)

#> $Global.statistics
#>   Dimensions  OR     P     g
#> 1      2 x 2 3.4 0.773 0.273
#> 

### 3 x 3 repeated matrix
data(AndersonRainGarden)
cohenG(AndersonRainGarden)

#> $Global.statistics
#>   Dimensions   OR    P    g
#> 1      3 x 3 11.5 0.92 0.42
#> 
#> $Pairwise.statistics
#>                                        Comparison  OR     P     g
#> 1       Yes.before/Yes.after : No.before/No.after Inf     1   0.5
#> 2 Yes.before/Yes.after : Maybe.before/Maybe.after  11 0.917 0.417
#> 3   No.before/No.after : Maybe.before/Maybe.after   7 0.875 0.375
#>