Compare fit statistics for glm models

Produces a table of fit statistics for multiple glm models.

compareGLM(fits, ...)

Arguments

fits
A series of model object names, separated by commas.

...
Other arguments passed to list.

Value

A list of two objects: The series of model calls, and a data frame of statistics for each model.

Details

Produces a table of fit statistics for multiple glm models: AIC, AICc, BIC, p-value, pseudo R-squared (McFadden, Cox and Snell, Nagelkerke).

Smaller values for AIC, AICc, and BIC indicate a better balance of goodness-of-fit of the model and the complexity of the model. The goal is to find a model that adequately explains the data without having too many terms.

BIC tends to choose models with fewer parameters relative to AIC. For comparisons with AIC, etc., to be valid, both models must have the same data, without transformations, use the same dependent variable, and be fit with the same method. They do not need to be nested.

References

http://rcompanion.org/rcompanion/e_07.html

See also

Examples

### Compare among logistic regresion models
data(AndersonBias)
model.0 = glm(Result ~ 1, weight = Count, data = AndersonBias,
             family = binomial(link="logit"))
model.1 = glm(Result ~ County, weight = Count, data = AndersonBias,
             family = binomial(link="logit"))
model.2 = glm(Result ~ County + Sex, weight = Count, data = AndersonBias,
             family = binomial(link="logit"))
model.3 = glm(Result ~ County + Sex + County:Sex, weight = Count,
             data = AndersonBias, family = binomial(link="logit"))
compareGLM(model.0, model.1, model.2, model.3)

#> $Models
#>   Formula                             
#> 1 "Result ~ 1"                        
#> 2 "Result ~ County"                   
#> 3 "Result ~ County + Sex"             
#> 4 "Result ~ County + Sex + County:Sex"
#> 
#> $Fit.criteria
#>   Rank Df.res   AIC  AICc   BIC McFadden Cox.and.Snell Nagelkerke  p.value
#> 1    1     15 254.8 255.7 256.3  0.00000        0.0000     0.0000      Inf
#> 2    4     12 255.8 261.8 259.7  0.01986        0.2675     0.2675 0.073780
#> 3    5     11 250.0 259.3 254.6  0.05101        0.5504     0.5504 0.005335
#> 4    8      8 248.8 278.8 255.7  0.07979        0.7137     0.7137 0.002153
#>